rafs: move v5 specific validate_digest() into v5.rs
Move v5 specific validate_digest() into v5.rs, so we could refactor
RafsInode later.
