+++ rafs/src/metadata/cached_v5.rs
@@ -20,8 +20,9 @@ use fuse_rs::abi::linux_abi;
 use fuse_rs::api::filesystem::Entry;
 
 use crate::metadata::layout::v5::{
-    rafsv5_alloc_bio_desc, RafsBlobEntry, RafsChunkFlags, RafsChunkInfo, RafsV5BlobTable,
-    RafsV5ChunkInfo, RafsV5Inode, RafsV5InodeFlags, RafsV5XAttrsTable, RAFSV5_ALIGNMENT,
+    rafsv5_alloc_bio_desc, rafsv5_validate_digest, RafsBlobEntry, RafsChunkFlags, RafsChunkInfo,
+    RafsV5BlobTable, RafsV5ChunkInfo, RafsV5Inode, RafsV5InodeFlags, RafsV5XAttrsTable,
+    RAFSV5_ALIGNMENT,
 };
 use crate::metadata::layout::{bytes_to_os_str, parse_xattr, RAFS_ROOT_INODE};
 use crate::metadata::{
@@ -30,6 +31,7 @@ use crate::metadata::{
 };
 use crate::RafsIoReader;
 
+use nydus_utils::digest::Algorithm;
 use nydus_utils::{digest::RafsDigest, ByteSize};
 
 pub struct CachedSuperBlockV5 {
@@ -137,6 +139,15 @@ impl RafsSuperInodes for CachedSuperBlockV5 {
         self.s_inodes
             .get(&ino)
             .map_or(Err(enoent!()), |i| Ok(i.clone()))
+    }
+
+    fn validate_digest(
+        &self,
+        inode: Arc<dyn RafsInode>,
+        recursive: bool,
+        digester: Algorithm,
+    ) -> Result<bool> {
+        rafsv5_validate_digest(inode, recursive, digester)
     }
 }
